toolkit/components/passwordmgr/test/mochitest/test_autocomplete_highlight_non_login.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test form field autofill highlight</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="text/javascript" src="../../../satchel/test/satchel_common.js"></script>
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content"></div>
<pre id="test">
Highlighting autofill fields in the test form

Fields that have been filled in automatically are highlighted by the pseudo class `:autofill`.
This test ensures that this pseudo-class is not applied when clicking on an insecure warning or footer.

<template id="form1-template">
  <form id="form1" action="https://autocomplete">
    <input type="text" name="uname">
    <input type="password" name="pword">
    <button type="submit" name="submit">Submit</button>
  </form>
</template>
<script>
  const formTemplate = document.getElementById("form1-template");

  add_setup(async () => {
    await setStoredLoginsDuringTest(
      [location.origin, "http://autocomplete", null, "user1", "pass1"],
      [location.origin, "http://autocomplete", null, "user2", "pass2"]
    );
    listenForUnexpectedPopupShown();
  });


function closeCurrentTab() {
  runInParent(function cleanUpWindow() {
    let window = Services.wm.getMostRecentWindow("navigator:browser");
    window.gBrowser.removeTab(window.gBrowser.selectedTab);
  });
}

  add_named_task("field highlight on pw field autocomplete with insecure warning", async () => {
    const form = setContentForTask(formTemplate);
    const autofillResult = await formAutofillResult(form.id);
    is(autofillResult, "no_saved_logins", "form has not been filled due to no saved logins");

    // Press enter on insecure warning and check.
    form.pword.focus();
    await popupByArrowDown();
    synthesizeKey("KEY_ArrowDown"); // insecure warning
    synthesizeKey("KEY_Enter");

    is(document.defaultView.getComputedStyle(form.pword).getPropertyValue("filter"), "none",
       "Highlight is not applied to the password field if enter key is pressed on the insecure warning item");
    is(document.defaultView.getComputedStyle(form.uname).getPropertyValue("filter"), "none",
       "Highlight is not applied to the username field if enter key is pressed on the insecure warning item");

    // Press tab on insecure warning and check.
    await openPopupOn(form.pword);
    synthesizeKey("KEY_ArrowDown"); // insecure warning
    synthesizeKey("KEY_Tab");

    is(document.defaultView.getComputedStyle(form.pword).getPropertyValue("filter"), "none",
       "Highlight is not applied to the password field if tab key is pressed on the insecure warning item");
    is(document.defaultView.getComputedStyle(form.uname).getPropertyValue("filter"), "none",
       "Highlight is not applied to the username field if tab key is pressed on the insecure warning item");
  });

  add_named_task("field highlight on pw field autocomplete footer", async () => {
    const form = setContentForTask(formTemplate);
    const autofillResult = await formAutofillResult(form.id);
    is(autofillResult, "no_saved_logins", "form has not been filled due to no saved logins");

    // Press enter on the footer and check.
    await openPopupOn(form.pword);
    synthesizeKey("KEY_ArrowUp"); // footer
    synthesizeKey("KEY_Enter");

    is(document.defaultView.getComputedStyle(form.pword).getPropertyValue("filter"), "none",
       "Highlight is not applied to the password field if enter key is pressed on the footer item");
    is(document.defaultView.getComputedStyle(form.uname).getPropertyValue("filter"), "none",
       "Highlight is not applied to the username field if enter key is pressed on the footer item");

    closeCurrentTab();

    // Press tab on the footer and check.
    await openPopupOn(form.pword);
    synthesizeKey("KEY_ArrowUp"); // footer
    synthesizeKey("KEY_Tab");

    is(document.defaultView.getComputedStyle(form.pword).getPropertyValue("filter"), "none",
       "Highlight is not applied to the password field if tab key is pressed on the footer item");
    is(document.defaultView.getComputedStyle(form.uname).getPropertyValue("filter"), "none",
       "Highlight is not applied to the username field if tab key is pressed on the insecure warning item");

    closeCurrentTab();
  });
</script>
</body>
</html>
